# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Globics default service.",
    "version": "1.0.0",
    "title": "Globics API SYSTEM SERVICES"
  },
  "host": "",
  "basePath": "/api",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/users/resentActivationEmail": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "resent activation email",
        "operationId": "resentActivationEmail",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "email",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "mail@gmail.com"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/users/avatar": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Upload avatar",
        "operationId": "uploadAvatar",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "type": "string",
            "required": true,
            "description": "Bearer xxxxxxx"
          },
          {
            "name": "avatar",
            "in": "formData",
            "description": "file to upload",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {}
      }
    },
...
  },
...
}